Master_Thesis/requirements.txt
absl-py==2.1.0
accelerate==0.26.1
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.1
appdirs==1.4.4
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.2.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
black==24.1a1
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cachetools==5.4.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cmake==3.20.3
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1704278392174/work
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1702070282333/work
cycler==0.12.1
datasets==2.16.1
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dill==0.3.7
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
evaluate==0.4.1
exceptiongroup==1.2.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock @ file:///croot/filelock_1700591183607/work
flake8==7.0.0
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.41
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
google-auth==2.32.0
google-auth-oauthlib==1.0.0
grpcio==1.65.1
huggingface-hub==0.20.2
idna @ file:///croot/idna_1666125576474/work
imageio==2.33.1
importlib_metadata==8.0.0
importlib_resources==6.1.3
iniconfig==2.0.0
inquirerpy==0.3.4
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1708996548741/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.3.2
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1704727043329/work
kaleido==0.2.1
kiwisolver==1.4.5
lazy_loader==0.3
libcst==1.1.0
lightning-utilities==0.10.1
Markdown==3.6
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.7.5
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mccabe==0.7.0
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
moreorless==0.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
mypy==1.8.0
mypy-extensions==1.0.0
-e git+https://github.com/SHI-Labs/NATTEN@6948a26b360e3cc499d814b1646ace45928c2db0#egg=natten
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx @ file:///croot/networkx_1690561992265/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python-headless==4.9.0.80
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.0.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathspec==0.12.1
peft==0.7.1
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pfzy==0.3.4
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///croot/pillow_1696580024257/work
platformdirs==4.1.0
plotly==5.20.0
pluggy==1.3.0
pretrainedmodels==0.7.4
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==4.25.2
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyflakes==3.2.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytest==7.4.4
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
qudida==0.0.4
regex==2023.12.25
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==2.0.0
responses==0.18.0
rsa==4.9
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
segmentation-models-pytorch==0.3.3
sentry-sdk==1.39.2
setproctitle==1.3.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
stdlibs==2023.12.15
sympy @ file:///croot/sympy_1701397643339/work
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.2
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
torch==2.0.1
torchaudio==2.0.2
torchinfo==1.8.0
torchmetrics==1.3.0.post0
torchvision==0.15.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827257044/work
tqdm==4.66.1
trailrunner==1.4.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1704212992681/work
-e git+https://github.com/huggingface/transformers.git@da3c79b245afcce88f5db79ada10bf5b7c200ab1#egg=transformers
triton==2.0.0
typing-inspect==0.9.0
typing_extensions @ file:///croot/typing_extensions_1705599297034/work
tzdata==2023.4
ufmt==2.3.0
urllib3 @ file:///croot/urllib3_1698257533958/work
usort==1.0.7
wandb==0.16.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
